add dashboard route
