+++ api-nest/package.json
@@ -45,10 +45,12 @@
     "cache-manager": "4.1.0",
     "class-transformer": "0.5.1",
     "class-validator": "0.13.2",
+    "date-fns": "2.29.1",
     "prisma": "4.1.0",
     "reflect-metadata": "0.1.13",
     "rimraf": "3.0.2",
-    "rxjs": "7.5.6"
+    "rxjs": "7.5.6",
+    "undici": "5.8.0"
   },
   "devDependencies": {
     "@nestjs/cli": "9.0.0",
